backend/requirements.txt
django~=2.2.2
django-autoslug~=1.9.4
